+++ udata/core/metrics/models.py
@@ -1,12 +1,13 @@
+from mongoengine.fields import DictField
+
 from udata.api_fields import field
-from udata.mongo import db
 
 __all__ = ("WithMetrics",)
 
 
 class WithMetrics(object):
     metrics = field(
-        db.DictField(),
+        DictField(),
         readonly=True,
         auditable=False,
     )