+++ udata/core/spatial/api_fields.py
@@ -71,15 +71,19 @@ spatial_coverage_fields = api.model(
             geojson, allow_null=True, description="A multipolygon for the whole coverage"
         ),
         "zones": fields.Raw(
-            attribute=lambda s: marshal(s.zones, zone_fields)
-            if request.headers.get(FULL_OBJECTS_HEADER, False, bool)
-            else [str(z) for z in s.zones],
+            attribute=lambda s: (
+                marshal(s.zones, zone_fields)
+                if request.headers.get(FULL_OBJECTS_HEADER, False, bool)
+                else [str(z) for z in s.zones]
+            ),
             description="The covered zones identifiers (full GeoZone objects if `X-Get-Datasets-Full-Objects` is set, IDs of the zones otherwise)",
         ),
         "granularity": fields.Raw(
-            attribute=lambda s: {"id": s.granularity or "other", "name": s.granularity_label}
-            if request.headers.get(FULL_OBJECTS_HEADER, False, bool)
-            else s.granularity,
+            attribute=lambda s: (
+                {"id": s.granularity or "other", "name": s.granularity_label}
+                if request.headers.get(FULL_OBJECTS_HEADER, False, bool)
+                else s.granularity
+            ),
             default="other",
             description="The spatial/territorial granularity (full Granularity object if `X-Get-Datasets-Full-Objects` is set, ID of the granularity otherwise)",
         ),