+++ udata/routing.py
@@ -3,6 +3,7 @@ from uuid import UUID
 
 from bson import ObjectId
 from flask import redirect, request, url_for
+from mongoengine import Q
 from mongoengine.errors import InvalidQueryError, ValidationError
 from werkzeug.exceptions import NotFound
 from werkzeug.routing import BaseConverter, PathConverter
@@ -13,7 +14,7 @@ from udata.core.spatial.models import GeoZone
 from udata.core.visualizations.models import Chart
 from udata.features.notifications.models import Notification
 from udata.harvest.models import HarvestSource
-from udata.mongo import db
+from udata.mongo.slug_fields import SlugField
 from udata.uris import cdata_url, homepage_url
 
 
@@ -68,7 +69,7 @@ class ModelConverter(BaseConverter):
 
     @property
     def has_slug(self):
-        return hasattr(self.model, "slug") and isinstance(self.model.slug, db.SlugField)
+        return hasattr(self.model, "slug") and isinstance(self.model.slug, SlugField)
 
     @property
     def has_redirected_slug(self):
@@ -90,7 +91,7 @@ class ModelConverter(BaseConverter):
             pass
         try:
             quoted = self.quote(value)
-            query = db.Q(slug=value) | db.Q(slug=quoted)
+            query = Q(slug=value) | Q(slug=quoted)
             obj = self.model.objects(query).exclude(*self.get_excludes()).get()
         except (InvalidQueryError, self.model.DoesNotExist):
             # If the model doesn't have a slug or matching slug doesn't exist.